Pull CategoryBin and its tests into json_schema
I've deleted the manual deserialization code because we can use serde's derive
feature instead.
