+++ src/ipums_metadata_model.rs
@@ -45,6 +45,7 @@
 //! information such as variable and value labels.
 //!
 //!
+use crate::json_schema::CategoryBin;
 use crate::layout::LayoutVar;
 use crate::mderror::MdError;
 use std::fmt;
@@ -194,97 +195,6 @@ impl IpumsCategory {
         }
     }
 }
-#[derive(Clone, Debug)]
-pub enum CategoryBin {
-    LessThan { value: i64, label: String },
-    Range { low: i64, high: i64, label: String },
-    MoreThan { value: i64, label: String },
-}
-
-impl CategoryBin {
-    pub fn new(low: Option<i64>, high: Option<i64>, label: &str) -> Result<Self, MdError> {
-        match (low, high) {
-            (Some(low), Some(high)) if high < low => Err(MdError::Msg(format!(
-                "category_bins: a low of {} and high of {} do not satisfy low <= high",
-                low, high
-            ))),
-            (Some(low), Some(high)) => Ok(Self::Range {
-                low,
-                high,
-                label: label.to_owned(),
-            }),
-            (None, Some(high)) => Ok(Self::LessThan {
-                value: high,
-                label: label.to_owned(),
-            }),
-            (Some(low), None) => Ok(Self::MoreThan {
-                value: low,
-                label: label.to_owned(),
-            }),
-            (None, None) => Err(MdError::Msg(
-                "category_bins: must have low, high, or both set to some value".to_string(),
-            )),
-        }
-    }
-
-    pub fn within(&self, test_value: i64) -> bool {
-        match self {
-            Self::LessThan { value, .. } => test_value < *value,
-            Self::Range { low, high, .. } => test_value >= *low && test_value <= *high,
-            Self::MoreThan { value, .. } => test_value > *value,
-        }
-    }
-}
-
-impl TryFrom<serde_json::Value> for CategoryBin {
-    type Error = MdError;
-
-    fn try_from(value: serde_json::Value) -> Result<Self, Self::Error> {
-        use MdError::ParsingError;
-
-        let Some(label_value) = value.get("value_label") else {
-            return Err(ParsingError(
-                "category_bins: missing required field 'value_label'".to_string(),
-            ));
-        };
-
-        let label = match label_value.as_str() {
-            Some(label) => label,
-            None => {
-                return Err(ParsingError(
-                    "category_bins: field 'value_label' must be a string".to_string(),
-                ));
-            }
-        };
-
-        let low = match value.get("low") {
-            None => None,
-            Some(low_value) => match low_value.as_i64() {
-                None => {
-                    return Err(ParsingError(
-                        "category_bins: field 'low' must be an integer".to_string(),
-                    ));
-                }
-                Some(low) => Some(low),
-            },
-        };
-
-        let high = match value.get("high") {
-            None => None,
-            Some(high_value) => match high_value.as_i64() {
-                None => {
-                    return Err(ParsingError(
-                        "category_bins: field 'high' must be an integer".to_string(),
-                    ));
-                }
-                Some(high) => Some(high),
-            },
-        };
-
-        let bin = CategoryBin::new(low, high, label)?;
-        Ok(bin)
-    }
-}
 
 mod test {
     #[cfg(test)]
@@ -321,118 +231,5 @@ mod test {
 
         assert_eq!(cat2.label(), cat3.label());
         assert_eq!("second", cat3.label());
-    }
-
-    #[test]
-    fn test_category_bin_new_less_than() {
-        let bin = CategoryBin::new(None, Some(3), "less than 3")
-            .expect("expected Ok(CategoryBin::LessThan)");
-        assert!(matches!(bin, CategoryBin::LessThan { .. }))
-    }
-
-    #[test]
-    fn test_category_bin_new_more_than() {
-        let bin = CategoryBin::new(Some(3), None, "more than 3")
-            .expect("expected Ok(CategoryBin::MoreThan)");
-        assert!(matches!(bin, CategoryBin::MoreThan { .. }));
-    }
-
-    #[test]
-    fn test_category_bin_new_range() {
-        let bin = CategoryBin::new(Some(3), Some(5), "between 3 and 5")
-            .expect("expected Ok(CategoryBin::Range)");
-        assert!(matches!(bin, CategoryBin::Range { .. }));
-    }
-
-    #[test]
-    fn test_category_bin_new_no_boundaries_error() {
-        let result = CategoryBin::new(None, None, "no boundaries!");
-        assert!(
-            result.is_err(),
-            "it should be an error if neither low nor high is provided"
-        );
-    }
-
-    #[test]
-    fn test_category_bin_new_high_less_than_low_error() {
-        let result = CategoryBin::new(Some(10), Some(2), "that's not possible");
-        assert!(result.is_err(), "it should be an error if high < low");
-    }
-
-    #[test]
-    fn test_category_bin_try_from_json_value() {
-        let json_str = "{\"code\": 1,\n\
-                        \"value_label\": \"1 to 14 hours worked per week\",\n\
-                        \"low\": 1,\n\
-                        \"high\": 14}";
-
-        let value: serde_json::Value =
-            serde_json::from_str(json_str).expect("test fixture JSON is invalid");
-
-        let category_bin: CategoryBin = value.try_into().expect("parsing into CategoryBin failed");
-
-        match category_bin {
-            CategoryBin::Range { low, high, label } => {
-                assert_eq!(low, 1);
-                assert_eq!(high, 14);
-                assert_eq!(label, "1 to 14 hours worked per week");
-            }
-            _ => {
-                panic!("expected variant CategoryBin::Range because both low and high are present");
-            }
-        }
-    }
-
-    /// It's an error if the value_label field is missing from the input JSON.
-    #[test]
-    fn test_category_bin_try_from_json_value_label_required() {
-        let json_str = "{\"code\": 1,\n\
-                         \"low\": 1,\n\
-                         \"high\": 14}";
-        let value: serde_json::Value =
-            serde_json::from_str(json_str).expect("test fixture JSON is invalid");
-        let result: Result<CategoryBin, _> = value.try_into();
-        assert!(result.is_err());
-    }
-
-    /// It's an error if the value_label field is present but isn't a string.
-    #[test]
-    fn test_category_bin_try_from_json_value_is_string() {
-        let json_str = "{\"code\": 1,\n\
-                        \"value_label\": 2,\n\
-                        \"low\": 1,\n\
-                        \"high\": 14}";
-        let value: serde_json::Value =
-            serde_json::from_str(json_str).expect("test fixture JSON is invalid");
-        let result: Result<CategoryBin, _> = value.try_into();
-        assert!(result.is_err());
-    }
-
-    /// It's an error if the low field is present but isn't an integer.
-    #[test]
-    fn test_category_bin_try_from_json_low_is_integer() {
-        let json_str = "{\"code\": 1,\n\
-                        \"value_label\": \"1 to 14 hours worked per week\",\n\
-                        \"low\": \"a\",\n\
-                        \"high\": 14}";
-
-        let value: serde_json::Value =
-            serde_json::from_str(json_str).expect("test fixture JSON is invalid");
-        let result: Result<CategoryBin, _> = value.try_into();
-        assert!(result.is_err());
-    }
-
-    /// It's an error if the high field is present but isn't an integer.
-    #[test]
-    fn test_category_bin_try_from_json_high_is_integer() {
-        let json_str = "{\"code\": 1,\n\
-                        \"value_label\": \"1 to 14 hours worked per week\",\n\
-                        \"low\": 1,\n\
-                        \"high\": \"b\"}";
-
-        let value: serde_json::Value =
-            serde_json::from_str(json_str).expect("test fixture JSON is invalid");
-        let result: Result<CategoryBin, _> = value.try_into();
-        assert!(result.is_err());
     }
 }
